+++ src/model/BelongsToMany.ts
@@ -1,6 +1,5 @@
 import { Builder } from "../query/Builder.js";
-import { insertAndResolveKey, type PrimaryKeyColumn } from "./PrimaryKeyResolution.js";
-import { Schema } from "../schema/Schema.js";
+import { insertAndResolveKey, primaryKeyColumn, type PrimaryKeyColumn } from "./PrimaryKeyResolution.js";
 import { Collection } from "../support/Collection.js";
 import { shouldGeneratePrimaryKeyForColumn, snakeCase } from "../utils.js";
 import type { Model, ModelMassAssignmentInputWithout, ModelConstructor, ModelKey, PivotQueryBuilder, StripTablePrefix } from "./Model.js";
@@ -374,7 +373,7 @@ export class BelongsToMany<T extends Record<string, any> = Model, RelatedFixed e
   }
 
   protected async pivotPrimaryKeyColumn(primaryKey: string): Promise<PrimaryKeyColumn | null> {
-    return await Schema.getColumn(this.qualifiedPivotTable(), primaryKey, this.parent.getConnection());
+    return await primaryKeyColumn(this.parent.getConnection(), this.qualifiedPivotTable(), primaryKey);
   }
 
   protected async shouldAutoGeneratePivotPrimaryKey(primaryKey: string): Promise<boolean> {